requirements.txt
black==22.12.0
openai==0.26.1
python-dotenv==0.21.0
